# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BCA Entertainment Works ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List users
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/deactivate:
    patch:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: ["string", "null"]
        designation:
          type: ["string", "null"]
        dateOfJoining:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - department
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: string
        designation:
          type: string
        dateOfJoining:
          type: string
      required:
        - name
        - email
        - password
        - role
        - department
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        phone:
          type: string
        designation:
          type: string
        dateOfJoining:
          type: string
        isActive:
          type: boolean
...